7.pizza_redis_pipeline/second-attempt/risk_evaluator/requirements.txt
pymongo
redis
pandas
python-dotenv
